Removed ModalHeader #11
